split strftime into separate file
